refactored_v2_async/requirements.txt
asyncpg==0.31.0
asyncssh==2.14.2
sshtunnel==0.4.0

gspread-asyncio==1.9.0
gspread==5.10.0
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0

aiohttp==3.9.1
requests==2.31.0

aiocache==0.12.2
aiofiles==23.2.1

# Scheduling
schedule==1.2.1

python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

injector==0.21.0

pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-asyncio==0.23.2

mypy==1.7.1

black==23.12.1
flake8==7.0.0
isort==5.13.2
